Project1/backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
pymongo==4.10.1
python-dotenv==1.0.1
PyJWT==2.10.1
bcrypt==4.2.1
